+++ test/compile_errors/binary_schema_bad_tdim.zig
@@ -1,0 +1,7 @@
+const fits = @import("zigfitsio");
+
+comptime {
+    _ = fits.BinarySchema(&.{
+        .{ .name = "COORD", .tform = "4E", .tdim = &.{ 3, 2 } },
+    });
+}