+++ test/compile_errors/binary_schema_bad_tdisp.zig
@@ -1,0 +1,7 @@
+const fits = @import("zigfitsio");
+
+comptime {
+    _ = fits.BinarySchema(&.{
+        .{ .name = "COUNT", .tform = "1J", .tdisp = "Q8" },
+    });
+}