+++ test/compile_errors/binary_schema_bad_vla_tdim.zig
@@ -1,0 +1,7 @@
+const fits = @import("zigfitsio");
+
+comptime {
+    _ = fits.BinarySchema(&.{
+        .{ .name = "VALUES", .tform = "1PJ(4)", .tdim = &.{10} },
+    });
+}