+++ test/compile_errors/iterator_sentinel_slice.zig
@@ -1,0 +1,7 @@
+const fits = @import("zigfitsio");
+
+comptime {
+    _ = fits.Iterator(struct {
+        x: [:0]u8,
+    }, error{});
+}